修改了数据库

+++ blog/blog-core/src/main/resources/sql/user.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` INT NOT NULL AUTO_INCREMENT,
   `username` VARCHAR(50),
   `password` VARCHAR(100),
   `nickname` VARCHAR(50),
@@ -8,9 +8,9 @@ CREATE TABLE `user` (
   `avatar` VARCHAR(100),
   `create_time` DATETIME,
   `last_login_time` DATETIME,
-  `enabled` INT(2),
-  `account_non_expired` INT(2),
-  `account_non_locked` INT(2),
-  `credentials_non_expired` INT(2),
+  `enabled` TINYINT,
+  `account_non_expired` TINYINT,
+  `account_non_locked` TINYINT,
+  `credentials_non_expired` TINYINT,
   PRIMARY KEY (`id`)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8
